Well-ordered String
You know a password is well-ordered string. Well-ordered string means
that the order of the characters is in an alphabetical increasing order.
Like “abEm” is awell-ordered number. However, “abmE” is not a well-order
number. Given an input # that tells you also how many digits are in the
password, print all possible passwords.